Homework.Lesson 6. Task 2. New Method shifUpRepeatedly and shiftDownRepeatedly          Lesson 7. All

+++ src/qa05/lessons_homework/lesson_3/task_2/Task_2.java
@@ -8,11 +8,10 @@ public class Task_2 {
             array[i] = (int) (Math.random() * 100);
             System.out.println("Elenemt " + i + " of array = " + array[i]);
         }
-        int averageOfSum = 0;
+        double averageOfSum = 0;
         for (int i = 0; i < array.length; i++) {
             averageOfSum += array[i];
         }
-        
         averageOfSum = averageOfSum / array.length; // а если деление будет с остатком? получишь некорректное среднее
         System.out.println("Average = " + averageOfSum);
     }
